Make various network requests asynchronous (#65)
* Send various network requests asynchronously

Use workers to make the "send_typing", and "send_message" network
requests asynchronous.

* Download user avatars and server icons asynchronously

Introduce a new worker that downloads data from a URL and caches it to
files in the disk. It uses a new dependency, xxhash, to hash the URLs to
hexadecimal strings and use that as filenames for the files it saves.